## Description

Before the move to MQT Core, the DD Package was designed as a
header-only library. The way we had to handle terminal nodes via static
inline member variables lead to the project not compiling with MSVC.
Now that the DD package is part of MQT Core, there really is no need for
it to be header-only.

This PR makes a first step into that direction and resolves the ackward
handling of the various DD terminals.
> **Warning**
> This is a breaking change. The previous solution declared a
> ```c++
>  static mNode terminalNode;
>  constexpr static mNode* terminal{&terminalNode};
> ```
> This now becomes
> ```c++
> static mNode terminal;
> ```
> Thus, it might be necessary to replace occurences of `mNode::terminal`
with `&mNode::terminal`. (similarly for `vNode` and `dNode`)

For the first time in years, the change sin this PR allow the project to
be compiled by MSVC again and we can skip all the nasty workarounds for
that.
This PR also addresses some of the warnings raised by MSVC so that none
of the compilers should be producing any warnings.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

+++ src/ecc/Q7Steane.cpp
@@ -50,7 +50,7 @@ void Q7Steane::measureAndCorrectSingle(bool xSyndrome) {
     // K3: IIIUUUU
     for (std::size_t c = 0; c < controls.size(); c++) {
       for (std::size_t q = 0; q < ecc.nRedundantQubits; q++) {
-        if (((q + 1) & (1 << c)) != 0) {
+        if (((q + 1) & (static_cast<std::size_t>(1U) << c)) != 0) {
           const auto target = static_cast<Qubit>(i + nQubits * q);
           if (xSyndrome) {
             qcMapped->x(target, controls.at(c));
